+++ axum-extra/src/middleware/mod.rs
@@ -1,0 +1,6 @@
+//! Additional types for creating middleware.
+
+pub mod middleware_fn;
+
+#[doc(inline)]
+pub use self::middleware_fn::{from_fn, Next};