+++ axum-extra/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- None.
+- Add `middleware::from_fn` for creating middleware from async functions
 
 # 0.1.0 (02. December, 2021)